改进web客户端体验，增加节点上下线功能

+++ html/js/wit3d.js
@@ -23,13 +23,28 @@ function updateNodeList(result, nodesDIV) {
         }
         $(node).find('div[id^="nodeNameIP"]').text(result[n].name+" ("+result[n].address+")  " + platform)
         $(node).find('div[id^="nodeResource"]').text(resource)
+        imgStatus = $(node).find('img[id^="nodeStatus"]')
         if (result[n].state == 0) {
-            $(node).find('img[id^="nodeStatus"]').attr("src", "img/online.png")
+            $(imgStatus).attr("src", "img/online.png")
+            $(imgStatus).attr("title", "节点已上线，点击后可切换为离线状态。")
+            $(imgStatus).attr("request", "../nodes/"+result[n].name+"/_offline")
         } else if (result[n].state == 1) {
-            $(node).find('img[id^="nodeStatus"]').attr("src", "img/offline.png")
+            $(imgStatus).attr("src", "img/offline.png")
+            $(imgStatus).attr("title", "节点已离线，点击后可切换为上线状态。")
+            $(imgStatus).attr("request", "../nodes/"+result[n].name+"/_online")
         } else {
-            $(node).find('img[id^="nodeStatus"]').attr("src", "img/unknown.png")
+            $(imgStatus).attr("src", "img/unknown.png")
+            $(imgStatus).attr("title", "节点已失去联系，请检查网络连接或者节点本身的运行情况。")
         }
+        $(imgStatus).unbind("click").click(function(){
+            req = $(this).attr("request")
+            if (req.length > 0) {
+                $.ajax({
+                    url: req,
+                    type: "PUT"
+                });
+            }
+        });
     }
 }
 
@@ -44,7 +59,7 @@ function setNodeTimer() {
 }
 
 var jobRowString = '<td> <a href="tasks.html?jobid={id}&jobname={name}&update={update}">{name}</a> </td> \
-                    <td>{state}</td> \
+                    <td><img src="img/{state}" data-toggle="tooltip" title="{state-tip}" height="24px" width="24px"></td> \
                     <td>{tasks}</td> \
                     <td> \
                         <div class="progress" style="height:9px; width: 100%"> \
@@ -53,17 +68,26 @@ var jobRowString = '<td> <a href="tasks.html?jobid={id}&jobname={name}&update={u
                     </td> \
                     <td>{start}</td> \
                     <td>{finish}</td> \
-                    <td><img id="cancel-{id}" class="operation_img" data-toggle="tooltip" data-placement="top" title="取消执行" src="img/cancel.png" alt="取消" height="21px" width="21px"> \
-                        <img id="delete-{id}" class="operation_img" data-toggle="tooltip" data-placement="top" title="删除记录" src="img/delete.png" alt="删除" height="21px" width="21px"> \
+                    <td><img id="cancel-{id}" data-toggle="tooltip" data-placement="top" title="取消执行" src="img/{cancel}.svg" alt="取消" height="21px" width="21px"> \
+                        <img id="delete-{id}" data-toggle="tooltip" data-placement="top" title="删除记录" src="img/{delete}.svg" alt="删除" height="21px" width="21px"> \
                     </td>'
 
-function getJobState(v) {
+function getJobStateTip(v) {
     if (v == 0) return "等待"
     if (v == 1) return "执行"
     if (v == 2) return "挂起"
     if (v == 3) return "完成"
     if (v == 4) return "失败"
     if (v == 5) return "取消"
+}
+
+function getJobStateImg(v) {
+    if (v == 0) return "queued.svg"
+    if (v == 1) return "executing.png"
+    if (v == 2) return "halted.svg"
+    if (v == 3) return "completed.png"
+    if (v == 4) return "failed.svg"
+    if (v == 5) return "terminated.png"
 }
 
 function isActiveJobState(v) {
@@ -85,7 +109,8 @@ function updateJobList(result, jobTable) {
         row = jobRowString.replace(/{id}/g, result[n].id)
         row = row.replace(/{name}/g, result[n].name)
         row = row.replace("{update}", isActiveJobState(result[n].state))
-        row = row.replace("{state}", getJobState(result[n].state))
+        row = row.replace("{state}", getJobStateImg(result[n].state))
+        row = row.replace("{state-tip}", getJobStateTip(result[n].state))
         row = row.replace("{tasks}", result[n].tasks)
         if (result[n].state == 4 || result[n].state == 5)
             row = row.replace("{progress-status}", "bg-danger")
@@ -95,35 +120,46 @@ function updateJobList(result, jobTable) {
         if (result[n].hasOwnProperty("exec_time"))
             row = row.replace("{start}", result[n].exec_time.substring(5, 19))
         else
-            row = row.replace("{start}", "")
+            row = row.replace("{start}", "&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;")
         if (result[n].hasOwnProperty("finish_time"))
             row = row.replace("{finish}", result[n].finish_time.substring(5, 19))
         else
-            row = row.replace("{finish}", "")
+            row = row.replace("{finish}", "&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;")
+        if (isActiveJobState(result[n].state) == "yes") {
+            row = row.replace("{cancel}", "cancel")
+            row = row.replace("{delete}", "delete-grey")
+        } else {
+            row = row.replace("{cancel}", "cancel-grey")
+            row = row.replace("{delete}", "delete")
+        }
         $(tr).html(row)
 
-        $("#cancel-"+result[n].id).click(function(){
-            var id = $(this).attr("id").substring(7, $(this).attr("id").length)
-            $.ajax({
-                url: "../jobs/"+id+"/_terminate",
-                type: "PUT",
-                success: function(result) {
-                    $.get("../jobs", function(result){
-                        updateJobList(result, "#job-table")
-                    });
-                }
+        if (isActiveJobState(result[n].state) == "yes") {
+            $("#cancel-"+result[n].id).unbind("click").click(function(){
+                var id = $(this).attr("id").substring(7, $(this).attr("id").length)
+                $.ajax({
+                    url: "../jobs/"+id+"/_terminate",
+                    type: "PUT",
+                    success: function(result) {
+                        $.get("../jobs", function(result){
+                            updateJobList(result, "#job-table")
+                        });
+                    }
+                });
             });
-        });
-        $("#delete-"+result[n].id).click(function(){
-            var id = $(this).attr("id").substring(7, $(this).attr("id").length)
-            $.ajax({
-                url: "../jobs/"+id,
-                type: "DELETE",
-                success: function(result) {
-                    $("#job-"+id).remove()
-                }
+        } else {
+            $("#cancel-"+result[n].id).unbind("click")
+            $("#delete-"+result[n].id).unbind("click").click(function(){
+                var id = $(this).attr("id").substring(7, $(this).attr("id").length)
+                $.ajax({
+                    url: "../jobs/"+id,
+                    type: "DELETE",
+                    success: function(result) {
+                        $("#job-"+id).remove()
+                    }
+                });
             });
-        });
+        }
     }
 }
 
@@ -145,7 +181,7 @@ function getQueryString(name) {
 }
 
 var taskRowString ='<td>{number}</td><td>{name}</td> \
-                    <td><img src="img/{state}.png" data-toggle="tooltip" title="{state-tip}" height="24px" width="24px"></td> \
+                    <td><img src="img/{state}" data-toggle="tooltip" title="{state-tip}" height="24px" width="24px"></td> \
                     <td>{node}</td> \
                     <td><div class="progress" style="height:9px; width: 100%"> \
                             <div class="progress-bar {progress-status}" style="width:{progress}%"></div> \
@@ -157,14 +193,14 @@ var taskRowString ='<td>{number}</td><td>{name}</td> \
                     <td><button type="button" class="btn btn-primary btn-sm" data-toggle="modal" data-target="#taskLogModal" data-id="{id}">查看日志</button></td>'
 
 function getTaskState(v) {
-    if (v == 0) return "queued"
-    if (v == 1) return "scheduled"
-    if (v == 2) return "dispatching"
-    if (v == 3) return "executing"
-    if (v == 4) return "completed"
-    if (v == 5) return "failed"
-    if (v == 6) return "aborted"
-    if (v == 7) return "terminated"
+    if (v == 0) return "queued.svg"
+    if (v == 1) return "scheduled.png"
+    if (v == 2) return "dispatching.png"
+    if (v == 3) return "executing.png"
+    if (v == 4) return "completed.png"
+    if (v == 5) return "failed.svg"
+    if (v == 6) return "aborted.svg"
+    if (v == 7) return "terminated.png"
 }
 
 function getTaskStateTip(v) {
@@ -199,7 +235,7 @@ function updateTaskList(result, taskTable) {
         if (result[n].hasOwnProperty("node"))
             row = row.replace("{node}", result[n].node)
         else
-            row = row.replace("{node}", "")
+            row = row.replace("{node}", "&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;")
         if (result[n].state == 5 || result[n].state == 6 || result[n].state == 7)
             row = row.replace("{progress-status}", "bg-danger")
         else
@@ -208,15 +244,15 @@ function updateTaskList(result, taskTable) {
         if (result[n].hasOwnProperty("start_time"))
             row = row.replace("{start}", result[n].start_time.substring(5, 19))
         else
-            row = row.replace("{start}", "")
+            row = row.replace("{start}", "&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;")
         if (result[n].hasOwnProperty("finish_time"))
             row = row.replace("{finish}", result[n].finish_time.substring(5, 19))
         else
-            row = row.replace("{finish}", "")
+            row = row.replace("{finish}", "&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;")
         if (result[n].hasOwnProperty("error"))
             row = row.replace("{error}", "<img src=\"img/alert.png\" data-toggle=\"tooltip\" title=\""+result[n].error+"\" height=\"18px\" width=\"18px\">")
         else
-            row = row.replace("{error}", "")
+            row = row.replace("{error}", "&nbsp;")
         $(tr).html(row)
 
         $('#taskLogModal').on('show.bs.modal', function (event) {
